feat: [AD-008] moved components of bottom bar to one dir
